backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
google-cloud-aiplatform>=1.65.0
google-cloud-firestore>=2.17.0
googlemaps>=4.10.0
google-auth>=2.34.0
google-api-python-client>=2.143.0
langgraph>=0.2.28
langchain-google-vertexai>=2.0.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.5
langchain-mcp-adapters>=0.1.0
mcp[cli]>=1.6.0
httpx>=0.27.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
